requirements.txt
pandas
numpy
matplotlib
youtube_transcript_api
openai-whisper
yt-dlp
